requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.3.0
chromadb>=0.4.15
pydantic==2.5.0
pydantic[email]==2.5.0
requests==2.31.0
tiktoken==0.5.1
httpx>=0.25.0

